+++ packages/extension-contracts/src/index.ts
@@ -1,2 +1,4 @@
 export * from "./ids.js";
+export * from "./manifest-json-schema.js";
+export * from "./manifest.js";
 export * from "./versioning.js";